Move event handlers from `Processes` to corresponding `Handlers` folders
