+++ backend/src/Examples/ExampleApp.Examples.Services/Handlers/Projects/PublishEmployeeAssignedToProjectAssignmentNotificationEH.cs
@@ -3,13 +3,13 @@ using ExampleApp.Examples.Domain.Projects.Events;
 using LeanCode.Pipe;
 using MassTransit;
 
-namespace ExampleApp.Examples.Services.Processes.Projects;
+namespace ExampleApp.Examples.Services.Handlers.Projects;
 
-public class PublishEmployeeAssignedToProjectAssignmentNotification : IConsumer<EmployeeAssignedToAssignment>
+public class PublishEmployeeAssignedToProjectAssignmentNotificationEH : IConsumer<EmployeeAssignedToAssignment>
 {
     private readonly ILeanPipePublisher<EmployeeAssignmentsTopic> topicPublisher;
 
-    public PublishEmployeeAssignedToProjectAssignmentNotification(
+    public PublishEmployeeAssignedToProjectAssignmentNotificationEH(
         ILeanPipePublisher<EmployeeAssignmentsTopic> topicPublisher
     )
     {
