+++ backend/src/Examples/ExampleApp.Examples.Services/Handlers/Projects/PublishEmployeeUnassignedFromAssignmentNotificationEH.cs
@@ -3,29 +3,25 @@ using ExampleApp.Examples.Domain.Projects.Events;
 using LeanCode.Pipe;
 using MassTransit;
 
-namespace ExampleApp.Examples.Services.Processes.Projects;
+namespace ExampleApp.Examples.Services.Handlers.Projects;
 
-public class PublishEmployeeAssignedToAssignmentNotification : IConsumer<EmployeeAssignedToAssignment>
+public class PublishEmployeeUnassignedFromAssignmentNotificationEH : IConsumer<EmployeeUnassignedFromAssignment>
 {
     private readonly ILeanPipePublisher<ProjectEmployeesAssignmentsTopic> topicPublisher;
 
-    public PublishEmployeeAssignedToAssignmentNotification(
+    public PublishEmployeeUnassignedFromAssignmentNotificationEH(
         ILeanPipePublisher<ProjectEmployeesAssignmentsTopic> topicPublisher
     )
     {
         this.topicPublisher = topicPublisher;
     }
 
-    public Task Consume(ConsumeContext<EmployeeAssignedToAssignment> context)
+    public Task Consume(ConsumeContext<EmployeeUnassignedFromAssignment> context)
     {
         var msg = context.Message;
 
         var topic = new ProjectEmployeesAssignmentsTopic { ProjectId = msg.ProjectId };
-        var notification = new EmployeeAssignedToAssignmentDTO
-        {
-            AssignmentId = msg.AssignmentId,
-            EmployeeId = msg.EmployeeId,
-        };
+        var notification = new EmployeeUnassignedFromAssignmentDTO { AssignmentId = msg.AssignmentId };
 
         return topicPublisher.PublishAsync(topic, notification, context.CancellationToken);
     }
